KAFKA-15351: Ensure log-start-offset not updated to local-log-start-offset when tiered storage is enabled on the topic.
Addressed the review comments.

Start from (k0, v0) instead of (k1, v1) to match with the offset.

Addressed the review comments.

support to override bootstrap_servers if required.

Reinitialize the clients only when all the brokers are down.

Enable OffloadAndConsumeFromLeader test with broker bounce.

KAFKA-15351: Ensure log-start-offset not updated to local-log-start-offset when tiered storage is enabled on the topic.