KAFKA-15756: [1/2] Migrate existing integration tests to run old protocol in new coordinator (#14781)
This patch updates the testing framework to support running tests with kraft and the new group coordinator introduced in the context of KIP-848. This can be done by using `kraft+kip-848` as a quorum. Note that this is temporary until we make it the default and only option in 4.0. To verify this, this patch also enables kraft and kraft+kip-848 in PlaintextConsumerTest and its parent classes.

Reviewers: David Jacot <djacot@confluent.io>