KAFKA-15931: Reopen TransactionIndex if channel is closed
refactor: move channel access into TransactionIndexFile

Reduce access to channel by exposing read/write/position on the TransactionIndexFile.

refactor: switch from File to Path when possible

Simplify code by using Path directly on TransactionIndexFile

Reduce access to TransactionIndexFile.

refactor: allow txn index to be read after closed

It is responsibility of the client to close the transaction index, but the transaction index should allow to read values even after closing.
The channel is an internal detail and should not affect reading values. If the file is available, then it should be read.
To encapsulate this dependency between file and buffer an internal abstraction TransactionIndexFile is introduced to handle the coordination between these two values.

refactor: collecting aborted transactions from remote files

Entry always contains txn index. If not existing on source, then it's an empty file on the cached entry.
Then, with this assumption, the logic to collect aborted transactions can be simplified.

Revert "KAFKA-15931: Reopen TransactionIndex if channel is closed"

This reverts commit 250c2c2222ce243313613daebd708dfa7ff770d2.

Revert "fixup! KAFKA-15931: Reopen TransactionIndex if channel is closed"

This reverts commit ed29d4da4c7ec7b60d30a692048afe5a215822fc.

Refactor search for aborted txns

KAFKA-15931: Reopen TransactionIndex if channel is closed

Cached TransactionIndex may get closed if interrupted, causing following calls to always fail with ClosedChannelException, and forcing process to be restarted.
In order to avoid this issue, a new method is exposed by TransactionIndex to validate state of channel; and index is reopened if closed.